+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * Registration module: standardize interpretation of the timepoint range "From" and "To" parameters. Now "From" and "To" correspond to zero-based time frame indices and the timepoint range include both "From" and "To" values.
+* Registration module: when using StackReg, round registration shifts to nearest integer.
 
 ### Removed